x86: Factor out 32-bit specific parts of the fastpath
